+++ core/Cargo.toml
@@ -55,11 +55,3 @@ indicatif = "0.17.0-beta.1"
 libsecp256k1 = "0.7.0"
 variant_count = "1.1.0"
 num_cpus = "1.13.0"
-
-[[bench]]
-name = "performator"
-harness = false
-
-[[bench]]
-name = "throughput"
-harness = false
