+++ core/src/json_interface.rs
@@ -8,13 +8,25 @@ where
 {
     match intermediate {
         Ok(t) => json!({
-        "tag": "Ok",
-        "content": t
+            "tag": "Ok",
+            "content": t
         }),
         Err(e) => json!({
-        "tag": "Err",
-        "content": e
+            "tag": "Err",
+            "content": e
         }),
     }
     .to_string()
+}
+
+#[cfg(test)]
+mod unit_tests {
+    use crate::json_interface::translate;
+    use crate::UnexpectedError;
+
+    #[test]
+    fn sanity_check() {
+        let a: Result<(), UnexpectedError> = Err(UnexpectedError("test".to_string()));
+        println!("{}", translate(a));
+    }
 }
