+++ core/src/model/errors.rs
@@ -1,0 +1,105 @@
+use crate::service::client::ApiError;
+use serde::ser::SerializeStruct;
+use serde::{Serialize, Serializer};
+use std::fmt::{Display, Formatter};
+use std::io::ErrorKind;
+
+#[derive(Debug)]
+pub struct UnexpectedError(pub String);
+
+impl Display for UnexpectedError {
+    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
+        write!(f, "unexpected error: {}", self.0)
+    }
+}
+
+impl Serialize for UnexpectedError {
+    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
+    where
+        S: Serializer,
+    {
+        let mut state = serializer.serialize_struct("UnexpectedError", 2)?;
+        state.serialize_field("tag", "Unexpected")?;
+        state.serialize_field("content", &self.0)?;
+        state.end()
+    }
+}
+
+#[derive(Debug, Serialize)]
+#[serde(tag = "tag", content = "content")]
+pub enum Error<U: Serialize> {
+    UiError(U),
+    Unexpected(String),
+}
+
+#[macro_export]
+macro_rules! unexpected {
+    ($base:literal $(, $args:tt )*) => {
+        Error::Unexpected(format!($base $(, $args )*))
+    };
+}
+
+#[macro_export]
+macro_rules! unexpected_only {
+    ($base:literal $(, $args:tt )*) => {
+        UnexpectedError(format!($base $(, $args )*))
+    };
+}
+
+#[derive(Debug, Clone, PartialEq)]
+pub enum CoreError {
+    AccountExists,
+    AccountNonexistent,
+    AccountStringCorrupted,
+    ClientUpdateRequired,
+    ClientWipeRequired,
+    DiskPathInvalid,
+    DiskPathTaken,
+    DrawingInvalid,
+    FileExists,
+    FileNameContainsSlash,
+    FileNameEmpty,
+    FileNonexistent,
+    FileNotDocument,
+    FileNotFolder,
+    FileParentNonexistent,
+    FolderMovedIntoSelf,
+    PathContainsEmptyFileName,
+    PathNonexistent,
+    PathStartsWithNonRoot,
+    PathTaken,
+    ImportCollision(String),
+    RootModificationInvalid,
+    RootNonexistent,
+    ServerUnreachable,
+    UsernameInvalid,
+    UsernamePublicKeyMismatch,
+    UsernameTaken,
+    Unexpected(String),
+}
+
+pub fn core_err_unexpected<T: std::fmt::Debug>(err: T) -> CoreError {
+    CoreError::Unexpected(format!("{:#?}", err))
+}
+
+impl From<std::io::Error> for CoreError {
+    fn from(e: std::io::Error) -> Self {
+        match e.kind() {
+            ErrorKind::NotFound | ErrorKind::PermissionDenied | ErrorKind::InvalidInput => {
+                CoreError::DiskPathInvalid
+            }
+            ErrorKind::AlreadyExists => CoreError::DiskPathTaken,
+            _ => core_err_unexpected(e),
+        }
+    }
+}
+
+impl<T: std::fmt::Debug> From<ApiError<T>> for CoreError {
+    fn from(e: ApiError<T>) -> Self {
+        match e {
+            ApiError::SendFailed(_) => CoreError::ServerUnreachable,
+            ApiError::ClientUpdateRequired => CoreError::ClientUpdateRequired,
+            e => core_err_unexpected(e),
+        }
+    }
+}
