+++ core/src/model/mod.rs
@@ -1,4 +1,5 @@
 pub mod client_conversion;
+pub mod errors;
 pub mod filename;
 pub mod repo;
 pub mod state;
