+++ core/src/repo/account_repo.rs
@@ -1,6 +1,6 @@
+use crate::model::errors::{core_err_unexpected, CoreError};
 use crate::model::state::Config;
 use crate::repo::local_storage;
-use crate::{core_err_unexpected, CoreError};
 use lockbook_models::account::{Account, ApiUrl};
 
 static ACCOUNT: &str = "ACCOUNT";
