+++ core/src/repo/local_storage.rs
@@ -1,4 +1,4 @@
-use crate::core_err_unexpected;
+use crate::model::errors::core_err_unexpected;
 use crate::model::state::Config;
 use crate::CoreError;
 use std::fs::{self, create_dir_all, read_dir, remove_file, File, OpenOptions};
