+++ core/src/service/mod.rs
@@ -1,4 +1,5 @@
 pub mod account_service;
+pub mod client;
 pub mod db_state_service;
 pub mod drawing_service;
 pub mod file_compression_service;
