+++ core/tests/change_document_content_tests.rs
@@ -3,8 +3,8 @@ mod integration_test;
 #[cfg(test)]
 mod change_document_content_tests {
     use lockbook_core::assert_matches;
-    use lockbook_core::client;
-    use lockbook_core::client::ApiError;
+    use lockbook_core::service::client;
+    use lockbook_core::service::client::ApiError;
     use lockbook_core::service::test_utils::{
         aes_encrypt, generate_account, generate_file_metadata, generate_root_metadata,
     };
