+++ core/tests/create_account_tests.rs
@@ -2,7 +2,7 @@ mod integration_test;
 
 #[cfg(test)]
 mod change_document_content_tests {
-    use lockbook_core::client;
+    use lockbook_core::service::client;
     use lockbook_core::service::test_utils::{generate_account, generate_root_metadata};
     use lockbook_models::api::*;
