+++ core/tests/create_file_tests.rs
@@ -4,8 +4,8 @@ mod integration_test;
 mod create_document_tests {
     use lockbook_core::assert_get_updates_required;
     use lockbook_core::assert_matches;
-    use lockbook_core::client;
-    use lockbook_core::client::ApiError;
+    use lockbook_core::service::client;
+    use lockbook_core::service::client::ApiError;
     use lockbook_core::service::test_utils::{
         generate_account, generate_file_metadata, generate_root_metadata,
     };
