+++ core/tests/get_document_tests.rs
@@ -3,8 +3,8 @@ mod integration_test;
 #[cfg(test)]
 mod get_document_tests {
     use lockbook_core::assert_matches;
-    use lockbook_core::client;
-    use lockbook_core::client::ApiError;
+    use lockbook_core::service::client;
+    use lockbook_core::service::client::ApiError;
     use lockbook_core::service::test_utils::{
         generate_account, generate_file_metadata, generate_root_metadata,
     };
