+++ core/tests/new_account_tests.rs
@@ -3,8 +3,8 @@ mod integration_test;
 #[cfg(test)]
 mod new_account_tests {
     use lockbook_core::assert_matches;
-    use lockbook_core::client;
-    use lockbook_core::client::ApiError;
+    use lockbook_core::service::client;
+    use lockbook_core::service::client::ApiError;
     use lockbook_core::service::test_utils::{generate_account, generate_root_metadata};
     use lockbook_models::api::*;
